Allow repeated keys in JSON objects
Chef Server 11 is returning repeated keys in the cookbook json
documents. The recipes object may contain repeated elements, so a custom
MapTypeAdapter has been registered to keep only the last one.

According to http://fadefade.com/json-comments.html, duplicate keys are
often used in json to emulate comments, and the standard behavior in
parsers is to keep the last key.